Add files via upload
remember the joke when someone said that their birthday is on 29th February and they only celebrate the birthday every four year, well it's not a joke anymore and that thing called The Leap Year. This is the script that can display whether this year is a leap year or not.